refactor: use top level statements in Program.cs
